Refactor out check command into multiple files
